Moved PostnatalCareFormsCcsRecordAggregationHelper to own file
