+++ custom/icds_reports/models/aggregate.py
@@ -18,7 +18,6 @@ from custom.icds_reports.utils.aggregation import (
     DailyFeedingFormsChildHealthAggregationHelper,
     DeliveryFormsAggregationHelper, GrowthMonitoringFormsAggregationHelper,
     InactiveAwwsAggregationHelper,
-    PostnatalCareFormsCcsRecordAggregationHelper,
     THRFormsCcsRecordAggregationHelper, THRFormsChildHealthAggregationHelper,
     AwwIncentiveAggregationHelper
 )
@@ -29,6 +28,8 @@ from six.moves import range
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms import ComplementaryFormsAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.complementary_forms_ccs_record import \
     ComplementaryFormsCcsRecordAggregationHelper
+from custom.icds_reports.utils.aggregation_helpers.postnatal_care_forms_ccs_record import \
+    PostnatalCareFormsCcsRecordAggregationHelper
 from custom.icds_reports.utils.aggregation_helpers.postnatal_care_forms_child_health import \
     PostnatalCareFormsChildHealthAggregationHelper
